projects/capstone/starter/backend/requirement.txt
alembic==1.7.7
aniso8601==6.0.0
astroid==2.2.5
attrs==22.1.0
Authlib==1.0.1
Babel==2.9.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
Click==7.0
cryptography==37.0.4
distlib==0.3.4
ecdsa==0.13.2
filelock==3.6.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.0
Flask-WTF==0.14.3
future==0.17.1
greenlet==1.1.2
gunicorn==20.0.4
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
isort==4.3.18
itsdangerous==2.0.1
Jinja2==2.10.1
lazy-object-proxy==1.4.0
Mako==1.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.5.2
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.11.0
pycparser==2.21
pycryptodome==3.3.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==3.0.9
pytest==6.2.2
python-dateutil==2.6.0
python-dotenv==0.20.0
python-jose-cryptodome==1.3.2
pytz==2019.1
requests==2.28.1
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.7
six==1.12.0
SQLAlchemy==1.3.4
toml==0.10.2
typed-ast==1.4.2
typing_extensions==4.2.0
urllib3==1.26.11
virtualenv==20.14.1
Werkzeug==2.0.3
wrapt==1.11.1
WTForms==3.0.1
zipp==3.8.0
